T31 feat: proninp - перенес решение в директорию Afisha, рефакторинг
