+++ Afisha/src/Afisha.Application/Abstractions/IUserSomeActionService.cs
@@ -1,0 +1,7 @@
+﻿using Afisha.Domain.Entities;
+
+namespace Afisha.Application.Abstractions;
+public interface IUserSomeActionService
+{
+    Task<User> SomeActionAsync();
+}
