+++ Afisha/src/Afisha.Application/Afisha.Application.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Afisha.Domain\Afisha.Domain.csproj"/>
+        <ProjectReference Include="..\Afisha.Domain\Afisha.Domain.csproj" />
     </ItemGroup>
 
 </Project>
