+++ Afisha/src/Afisha.Application/Contracts/Repositories/IRepository.cs
