+++ Afisha/src/Afisha.Application/Contracts/Repositories/IUserRepository.cs
