+++ Afisha/src/Afisha.Domain/Afisha.Domain.csproj
