+++ Afisha/src/Afisha.Domain/Contracts/ILocationService.cs
