+++ Afisha/src/Afisha.Domain/Entities/Abstractions/EntityBase.cs
