+++ Afisha/src/Afisha.Domain/Entities/Event.cs
