+++ Afisha/src/Afisha.Domain/Entities/Location.cs
