+++ Afisha/src/Afisha.Domain/Entities/LocationEvent.cs
