+++ Afisha/src/Afisha.Infrastructure/Data/Configurations/EventConfiguration.cs
