+++ Afisha/src/Afisha.Infrastructure/Data/Configurations/LocationEventConfiguration.cs
