+++ Afisha/src/Afisha.Infrastructure/Data/Repositories/UserRepository.cs
@@ -13,7 +13,7 @@ namespace Afisha.Infrastructure.Repositories
         public async Task<User?> AddUserAsync(User user, CancellationToken cancellationToken)
         {
             // Поиск логина в бд , для создания пользователя нужен уникальный логин (?)
-            var result = await context.Users.FirstOrDefaultAsync(x => x.UserName == user.UserName);
+            var result = await context.Users.FirstOrDefaultAsync(x => x.Login == user.Login);
             if (result is not null)
             {
                 throw new Exception("Пользователь с таким логином уже существует");
