--- src/Afisha.Domain/Contracts/IUserSomeActionService.cs
@@ -1,9 +1,0 @@
-﻿using Afisha.Domain.Entities;
-
-namespace Afisha.Domain.Contracts
-{
-    public interface IUserSomeActionService
-    {
-        Task<User> SomeActionAsync();
-    }
-}
