--- src/Afisha.Infrastructure/Services/UserNotificationSender.cs
@@ -1,13 +1,0 @@
-﻿using Afisha.Application.Contracts;
-using Afisha.Domain.Entities;
-
-namespace Afisha.Application.Services
-{
-    public class UserNotificationSender : IUserNotificationSender
-    {
-        public Task SendNotificationAsync(User to, User from, string subject, string body)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
